requirements.txt
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nmap==0.7.1
python-telegram-bot==22.2
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrdet==2.5
qreader==3.16
quadrilateral-fitter==1.12
requests==2.32.4
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
urllib3==2.5.0
